started writing fit_INLA routine
